#1535 generating test artifacts with a specified size for Maven

+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-npm-layout-provider/src/test/java/org/carlspring/strongbox/testing/artifact/NpmArtifactGeneratorStrategy.java
@@ -19,7 +19,7 @@ public class NpmArtifactGeneratorStrategy implements ArtifactGeneratorStrategy<N
     public Path generateArtifact(NpmArtifactGenerator artifactGenerator,
                                  String id,
                                  String version,
-                                 int size,
+                                 long size,
                                  Map<String, Object> attributesMap)
             throws IOException
     {
